Revert "Merge pull request #1 from Teasty/Pieces"
This reverts commit 027196ab2bf40db442bc32f9bdf7439ee01dde01, reversing
changes made to 99083d4fd0e786631ac4e1a3788e648cfcf595c3.
