--- Cube/.vs/Cube/xs/UserPrefs.xml
@@ -1,16 +1,0 @@
-﻿<Properties StartupConfiguration="{B372FD90-C149-214E-0CD2-737E9BF63DB7}|">
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/Scripts/Piece.cs">
-    <Files>
-      <File FileName="Assets/Scripts/RotationScript.cs" Line="39" Column="1" />
-      <File FileName="Assets/Scripts/SwipeController.cs" Line="1" Column="1" />
-      <File FileName="Assets/Scripts/Piece.cs" Line="1" Column="1" />
-    </Files>
-  </MonoDevelop.Ide.Workbench>
-  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
-  <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore />
-  </MonoDevelop.Ide.DebuggingService.Breakpoints>
-  <MonoDevelop.Ide.DebuggingService.PinnedWatches />
-  <MonoDevelop.Ide.ItemProperties.Cube PreferredExecutionTarget="Unity.Editor" />
-  <MultiItemStartupConfigurations />
-</Properties>
