--- Cube/Assets/Materials/Red.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 0cb2659c5952b4f669f1cbb2c1e36319
-timeCreated: 1521816016
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
