--- Cube/Assets/Prefubs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 39d37035c798343ea89eb7b90eea4ded
-folderAsset: yes
-timeCreated: 1521911333
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
