--- Cube/Assets/Prefubs/Cube.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 9119203ff4e4c40cca10ef6ce360f33e
-timeCreated: 1521911350
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
