--- Cube/Assets/Prefubs/EndPiece.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 8f2a4180263014c3f96734724d63dbec
-timeCreated: 1521911345
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
