--- Cube/Assets/Prefubs/GameObject.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: a87cd18942c68402ca66bf6cdfd1ca38
-timeCreated: 1521911353
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
