--- Cube/Assets/Prefubs/IPiece.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 0fc56ffeebc1348e69dee6266d0f1039
-timeCreated: 1521911342
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
