--- Cube/Assets/Prefubs/Main Camera.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: fe62f749b9d90488281a160bc2448eaa
-timeCreated: 1521911367
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
