--- Cube/Assets/Prefubs/StartPiece.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: d637f64d316604bbb91cce93144e2212
-timeCreated: 1521911338
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
