--- Cube/Assets/Scripts/EndPiece.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 29095e8179a1d42ccbcfbbdab0cb2b66
-timeCreated: 1521909486
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
