--- Cube/Assets/Scripts/NewPiece.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 7af599b2ce58b4a8e97370a8db6ea925
-timeCreated: 1521909141
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
