--- Cube/Assets/Scripts/StartPiece.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 012d89ec550244e32927d5e50f04e9df
-timeCreated: 1521908513
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
