+++ Cube/Assets/Scripts/SwipeController.cs
@@ -8,14 +8,6 @@ public class SwipeController : MonoBehaviour {
 	private bool isSwiping = false;
 	private Vector2 startTouch, swipeDelta;
 
-
-
-    void Start()
-    {
-        Screen.orientation = ScreenOrientation.Portrait;
-    }
-
-
 	void Update () {
 		tap = swipeLeft = swipeUp = swipeRight = swipeDown = false;
