+++ Cube/ProjectSettings/TagManager.asset
@@ -3,8 +3,7 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags:
-  - Piece
+  tags: []
   layers:
   - Default
   - TransparentFX
